Move `BuildStep` and metric logging into `build_helper`
